requirements.txt
python-dotenv==0.19.2
beautifulsoup4==4.10.0